Move RhoCmd to its own file
Overall goal: get ready to import JBoss Scanner modules by making our
code use different Ansible modules.

This commit: one small move towards modularization.
